added new well looking footer
